+++ packages/components/workbench/src/WorkbenchSidebar/WorkbenchSidebar.styles.ts
@@ -3,12 +3,13 @@ import tokens from '@contentful/f36-tokens';
 
 export const getWorkbenchSidebarStyles = (position = 'left') => {
   const borderStyleKey = position === 'left' ? 'borderRight' : 'borderLeft';
+  const width = position === 'left' ? '280px' : '360px';
 
   return {
     workbenchSidebar: css({
       backgroundColor: tokens.gray100,
       [borderStyleKey]: `1px solid ${tokens.gray200}`,
-      width: '280px',
+      width,
       padding: tokens.spacingL,
       overflowY: 'auto',
       overflowX: 'hidden',