+++ packages/core/src/Box/Box.tsx
@@ -92,6 +92,8 @@ function _Box<E extends React.ElementType = typeof BOX_DEFAULT_TAG>(
   );
 }
 
+_Box.displayName = 'Box';
+
 export const Box: PolymorphicComponent<
   ExpandProps<BoxInternalProps>,
   typeof BOX_DEFAULT_TAG