+++ packages/components/card/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.3.8](https://github.com/contentful/forma-36/compare/@contentful/f36-card@4.3.7...@contentful/f36-card@4.3.8) (2022-03-14)
+
+**Note:** Version bump only for package @contentful/f36-card
+
+
+
+
+
+## [4.3.7](https://github.com/contentful/forma-36/compare/@contentful/f36-card@4.3.6...@contentful/f36-card@4.3.7) (2022-03-10)
+
+**Note:** Version bump only for package @contentful/f36-card
+
+
+
+
+
+## [4.3.6](https://github.com/contentful/forma-36/compare/@contentful/f36-card@4.3.5...@contentful/f36-card@4.3.6) (2022-03-09)
+
+**Note:** Version bump only for package @contentful/f36-card
+
+
+
+
+
 ## [4.3.5](https://github.com/contentful/forma-36/compare/@contentful/f36-card@4.3.4...@contentful/f36-card@4.3.5) (2022-03-07)