+++ packages/components/entity-list/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.24](https://github.com/contentful/forma-36/compare/@contentful/f36-entity-list@4.1.23...@contentful/f36-entity-list@4.1.24) (2022-03-14)
+
+**Note:** Version bump only for package @contentful/f36-entity-list
+
+
+
+
+
+## [4.1.23](https://github.com/contentful/forma-36/compare/@contentful/f36-entity-list@4.1.22...@contentful/f36-entity-list@4.1.23) (2022-03-10)
+
+**Note:** Version bump only for package @contentful/f36-entity-list
+
+
+
+
+
+## [4.1.22](https://github.com/contentful/forma-36/compare/@contentful/f36-entity-list@4.1.21...@contentful/f36-entity-list@4.1.22) (2022-03-09)
+
+**Note:** Version bump only for package @contentful/f36-entity-list
+
+
+
+
+
 ## [4.1.21](https://github.com/contentful/forma-36/compare/@contentful/f36-entity-list@4.1.20...@contentful/f36-entity-list@4.1.21) (2022-03-03)