+++ packages/components/menu/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.9](https://github.com/contentful/forma-36/compare/@contentful/f36-menu@4.2.8...@contentful/f36-menu@4.2.9) (2022-03-14)
+
+**Note:** Version bump only for package @contentful/f36-menu
+
+
+
+
+
+## [4.2.8](https://github.com/contentful/forma-36/compare/@contentful/f36-menu@4.2.7...@contentful/f36-menu@4.2.8) (2022-03-10)
+
+**Note:** Version bump only for package @contentful/f36-menu
+
+
+
+
+
+## [4.2.7](https://github.com/contentful/forma-36/compare/@contentful/f36-menu@4.2.6...@contentful/f36-menu@4.2.7) (2022-03-09)
+
+**Note:** Version bump only for package @contentful/f36-menu
+
+
+
+
+
 ## [4.2.6](https://github.com/contentful/forma-36/compare/@contentful/f36-menu@4.2.5...@contentful/f36-menu@4.2.6) (2022-03-03)