+++ packages/components/note/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.19](https://github.com/contentful/forma-36/compare/@contentful/f36-note@4.2.18...@contentful/f36-note@4.2.19) (2022-03-14)
+
+**Note:** Version bump only for package @contentful/f36-note
+
+
+
+
+
+## [4.2.18](https://github.com/contentful/forma-36/compare/@contentful/f36-note@4.2.17...@contentful/f36-note@4.2.18) (2022-03-10)
+
+
+### Bug Fixes
+
+* **note:** allow rendering children in div ([#1938](https://github.com/contentful/forma-36/issues/1938)) ([23a4ece](https://github.com/contentful/forma-36/commit/23a4ece76cf774c45d42b7f3c0d7d148444a08de))
+
+
+
+
+
+## [4.2.17](https://github.com/contentful/forma-36/compare/@contentful/f36-note@4.2.16...@contentful/f36-note@4.2.17) (2022-03-10)
+
+**Note:** Version bump only for package @contentful/f36-note
+
+
+
+
+
+## [4.2.16](https://github.com/contentful/forma-36/compare/@contentful/f36-note@4.2.15...@contentful/f36-note@4.2.16) (2022-03-09)
+
+**Note:** Version bump only for package @contentful/f36-note
+
+
+
+
+
 ## [4.2.15](https://github.com/contentful/forma-36/compare/@contentful/f36-note@4.2.14...@contentful/f36-note@4.2.15) (2022-03-03)